+++ runtests.py
@@ -3,9 +3,37 @@ import sys
 import os
 from multiprocessing import current_process
 
+try:
+    import pep8
+except ImportError:
+    pep8 = None
+
 
 def run(**params):
     args = params.get('argv', sys.argv)
+    if '--pep8' in args:
+        args.remove('--pep8')
+        if pep8:
+            print('Running pep8.')
+            pep8style = pep8.StyleGuide(paths=['pulsar', 'examples'],
+                                        config_file='setup.cfg')
+            options = pep8style.options
+            report = pep8style.check_files()
+            if options.statistics:
+                report.print_statistics()
+            if options.benchmark:
+                report.print_benchmark()
+            if options.testsuite and not options.quiet:
+                report.print_results()
+            if report.total_errors:
+                if options.count:
+                    sys.stderr.write(str(report.total_errors) + '\n')
+                sys.exit(1)
+            print('OK')
+            sys.exit(0)
+        else:
+            print('pep8 must be installed')
+            sys.exit(1)
     if '--coverage' in args or params.get('coverage'):
         import coverage
         p = current_process()